Fixed the Spring implementation, began using Autowired beans, started work on the Thymeleaf portal
Spring dependency injection wasn't working with MessageController. After
several days of playing about with it I finally figured it out - in the
JDABuilder I was 'new MessageController' for the listener, which of course
stops beans from being wired into it.

Solution: Moved the onMessageReceived event listener into its own class,
MessageListener, and autowired this into the class constructing the
JDABuilder.