Refactored code that handles incoming trial commands into handlers
We now have an event listeners specifically for trial commands. The listener
will cycle through all beans which extend the TrialCommandHandler class and call
the handleCommandIfTriggered. This method checks if the command being called
should be handled by this handler and also if the number of arguments provided
are correct. If these conditions are both true then we call the handleCommand
method which should handle all things related to the command.

If we want the code to handle new trial commands in the future, then all that is
required is a new implementation of the TrialCommandHandler class.